Move sarif parser and tests, build completing
