+++ extensions/ql-vscode/src/vscode-tests/no-workspace/sarif/emptyResultsSarif.sarif
