+++ extensions/ql-vscode/src/vscode-tests/no-workspace/sarif/validSarif.sarif
